requirements.txt
blosc==1.11.1
catboost==1.2.5
contourpy==1.1.1
cycler==0.12.1
certifi
dill==0.3.8
dnspython==2.6.1
fonttools==4.53.0
graphviz==0.20.3
imbalanced-learn==0.12.3
imblearn==0.0
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
lightgbm==4.4.0
matplotlib==3.7.5
miceforest==5.7.0
numpy==1.24.4
pandas==2.0.3
pillow==10.3.0
plotly==5.22.0
pymongo==4.2.0
pyparsing==3.1.2
python-dotenv
pytz==2024.1
pywin32==306
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sensor==0.0.1
tenacity==8.3.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
pyYAML
-e .
